tests/wpt/web-platform-tests/shadow-dom/untriaged/user-interaction/ranges-and-selections/test-001.html

<!DOCTYPE html>
<html>
<head>
<title>Shadow DOM Test: A_07_01_01</title>
<link rel="author" title="Sergey G. Grekhov" href="mailto:sgrekhov@unipro.ru">
<link rel="help" href="http://www.w3.org/TR/2013/WD-shadow-dom-20130514/#ranges-and-selection">
<meta name="assert" content="User Interaction: Selection, returned by the window.getSelection() method must never return a selection within a shadow tree">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../../../../html/resources/common.js"></script>
<script src="../../../resources/shadow-dom-utils.js"></script>
</head>
<body>
<div id="log"></div>
<script>
test(unit(function (ctx) {
    var d = newRenderedHTMLDocument(ctx);

    var host = d.createElement('div');
    d.body.appendChild(host);
    var s = host.attachShadow({mode: 'open'});

    var span = d.createElement('span');
    span.innerHTML = 'Some text';
    s.appendChild(span);

    var range = d.createRange();
    range.setStart(span.firstChild, 0);
    range.setEnd(span.firstChild, 3);

    var selection = window.getSelection();
    selection.removeAllRanges();
    selection.addRange(range);

    var sl = window.getSelection();
    assert_equals(sl.toString(), '', 'window.getSelection() method must never return a selection ' +
            'within a shadow tree');

}), 'A_07_07_01_T01');


// test distributed nodes
test(unit(function (ctx) {
    var d = newRenderedHTMLDocument(ctx);

    var host = d.createElement('div');
    d.body.appendChild(host);

    var span = d.createElement('span');
    span.innerHTML = 'Some text';
    span.setAttribute('slot', 'span');
    host.appendChild(span);

    var s = host.attachShadow({mode: 'open'});
    s.innerHTML = '<slot name="span"></slot>';

    var range = d.createRange();
    range.setStart(span.firstChild, 0);
    range.setEnd(span.firstChild, 3);

    var selection = window.getSelection();
    selection.removeAllRanges();
    selection.addRange(range);

    var sl = window.getSelection();
    assert_equals(sl.toString(), '', 'window.getSelection() method must never return a selection ' +
        'within a shadow tree');

}), 'A_07_07_01_T02');
</script>
</body>
</html>
